Documenting local_commands, changing default cmds.
This patch adds documentation of the local_commands module.
It also changes the way local_commands works:

   If there is a local_commands module, the default commands are removed.

   If no cmd_help is defined, the default is used.

   If no cmd_fallback is defined, use the default cmd_g.

   Default commands are renamed to default_cmd_<name>.

Because of this, you can now create a local_commands module without having
to pick up the default commands if you don't want them.  If you do, it's a
simple matter of importing them from cony and removing the "default_"
prefix.